requirements.txt
Flask==2.3.3
supabase==2.3.4
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.1.2
Pillow==10.0.1
pytz==2024.1
openpyxl==3.1.2
requests==2.31.0